<2020.12.31 HeeSang-Yun> 1. 보상 지급 부분에서 값이 정상적으로 맞지 않는 버그 Fix

+++ cJSON.h
@@ -10,10 +10,10 @@ public:
 	~cJSON();
 
 	bool bRead_WalletInfo(const std::string _WalletInfo_Json, std::string &_MasterWalletAddr, std::string &_MasterWalletPrivateKey);
-	bool bRead_ResultInfo(const std::string _ResultInfo_Json, const std::string _Master_Wallet_Addr, const std::string _Master_Wallet_PrivateKey, float& _nTotalRewardCount);
+	bool bRead_ResultInfo(const std::string _ResultInfo_Json, const std::string _Master_Wallet_Addr, const std::string _Master_Wallet_PrivateKey, long double& _nTotalRewardCount);
 	bool bRead_RewardResultInfo(const std::string _RewardResultInfoJson);
 	
-	bool bWrite_RewardResult(std::map<int, StReward_Result> _ResultMap, std::string FaileName, float nTotalRewardCoin);
+	bool bWrite_RewardResult(std::map<int, StReward_Result> _ResultMap, std::string FaileName, long double nTotalRewardCoin);
 
 	struct stBalanceInfo Get_BalanceInfo(const std::string _String_Json);
 	__int64   Get_GasPrice(const std::string _String_Json);
